+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.1.0
+
+- Support for custom ellipsis
+- Allow more flexible usage through `charWidth`
+
 # 1.0.1
 
 - Fix hook return type TS definition